requirements.txt
Django==1.6
Pillow==2.5.1
South==1.0
argparse==1.2.1
django-autoslug==1.7.2
djangorestframework==2.3.14
easy-thumbnails==2.0.1
psycopg2==2.5.3
wsgiref==0.1.2
